ai-engine/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
httpx==0.27.0
python-dotenv==1.0.1

# NLP & ML
spacy==3.7.4
sentence-transformers==3.0.1
scikit-learn==1.5.0
numpy==1.26.4
pandas==2.2.2
joblib==1.4.2

# Resume parsing
pdfminer.six==20231228
python-docx==1.1.2
Pillow==10.3.0

# OpenAI
openai==1.30.1

# Utilities
python-multipart==0.0.9
aiofiles==23.2.1
